test-repositories/python-repo/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
werkzeug==3.0.1
pyjwt==2.8.0
python-dotenv==1.0.1
email-validator==2.1.0
pytest==7.4.0
pytest-flask==1.3.0 